+++ static/js/main.js
@@ -14,6 +14,7 @@ import * as reflect from "./views/reflect.js";
 import * as settings from "./views/settings.js";
 import * as session from "./views/session.js";
 import * as today from "./views/today.js";
+import * as inbox from "./views/inbox.js";
 
 register("capture", capture.render);
 register("history", history.render);
@@ -24,6 +25,7 @@ register("reflect", reflect.render);
 register("settings", settings.render);
 register("session", session.render);
 register("today", today.render);
+register("inbox", inbox.render);
 
 restorePrefs();
 mountShell();